feat(security): add security
add security, more unit tests and functionnals tests. Resolve bugs.

issue 7: https://github.com/alexandrecorroy/bileMo/issues/7

+++ src/UI/Action/CustomerUser/DeleteCustomerUserAction.php
@@ -16,17 +16,19 @@ namespace App\UI\Action\CustomerUser;
 use App\Repository\Interfaces\CustomerUserRepositoryInterface;
 use App\UI\Action\CustomerUser\Interfaces\DeleteCustomerUserActionInterface;
 use App\UI\Responder\CustomerUser\Interfaces\DeleteCustomerUserResponderInterface;
+use App\UI\Responder\CustomerUser\Interfaces\ForbiddenCustomerUserResponderInterface;
 use App\UI\Responder\CustomerUser\Interfaces\NotFoundCustomerUserResponderInterface;
 use Doctrine\Common\Cache\ApcuCache;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 /**
  * Class DeleteCustomerUserAction.
  *
- * @Route("/customerUser/{id}", name="customer_user_delete", methods={"DELETE"})
+ * @Route("api/customerUser/{id}", name="customer_user_delete", methods={"DELETE"})
  */
 final class DeleteCustomerUserAction implements DeleteCustomerUserActionInterface
 {
@@ -41,12 +43,21 @@ final class DeleteCustomerUserAction implements DeleteCustomerUserActionInterfac
     private $entityManager;
 
     /**
+     * @var TokenStorageInterface
+     */
+    private $tokenStorage;
+
+    /**
      * {@inheritdoc}
      */
-    public function __construct(EntityManagerInterface $entityManager, CustomerUserRepositoryInterface $customerUserRepository)
-    {
-        $this->entityManager = $entityManager;
+    public function __construct(
+        EntityManagerInterface $entityManager,
+        CustomerUserRepositoryInterface $customerUserRepository,
+        TokenStorageInterface $tokenStorage
+    ) {
+        $this->entityManager          = $entityManager;
         $this->customerUserRepository = $customerUserRepository;
+        $this->tokenStorage           = $tokenStorage;
     }
 
     /**
@@ -55,18 +66,22 @@ final class DeleteCustomerUserAction implements DeleteCustomerUserActionInterfac
     public function __invoke(
         Request $request,
         DeleteCustomerUserResponderInterface $deleteCustomerUserResponder,
-        NotFoundCustomerUserResponderInterface $notFoundCustomerUserResponder
+        NotFoundCustomerUserResponderInterface $notFoundCustomerUserResponder,
+        ForbiddenCustomerUserResponderInterface $forbiddenCustomerUserResponder
     ): Response {
 
         $cache = new ApcuCache();
 
         $customerUser = $this->customerUserRepository->findOneByUuidField($request->get("id"));
 
-        // check if Customer is own of customerUser
-
         if(\is_null($customerUser))
         {
             return $notFoundCustomerUserResponder();
+        }
+
+        if($customerUser->getCustomer()!==$this->tokenStorage->getToken()->getUser())
+        {
+            return $forbiddenCustomerUserResponder();
         }
 
         $cache->delete('find'.$customerUser->getUid());
